refactor: move styles to styles.go
