+++ tree/tree.go
@@ -13,75 +13,6 @@ import (
 	"github.com/charmbracelet/bubbles/viewport"
 )
 
-// StyleFunc allows the tree to be styled per item.
-type StyleFunc func(children Nodes, i int) lipgloss.Style
-
-// Styles contains style definitions for this tree component. By default, these
-// values are generated by DefaultStyles.
-type Styles struct {
-	TreeStyle lipgloss.Style
-	HelpStyle lipgloss.Style
-
-	selectedNodeFunc      StyleFunc
-	SelectedNodeStyle     lipgloss.Style
-	SelectedNodeStyleFunc StyleFunc
-
-	nodeFunc      StyleFunc
-	NodeStyle     lipgloss.Style
-	NodeStyleFunc StyleFunc
-
-	rootNodeFunc      StyleFunc
-	RootNodeStyle     lipgloss.Style
-	RootNodeStyleFunc StyleFunc
-
-	parentNodeFunc      StyleFunc
-	ParentNodeStyle     lipgloss.Style
-	ParentNodeStyleFunc StyleFunc
-
-	CursorStyle lipgloss.Style
-
-	EnumeratorStyle    lipgloss.Style
-	OpenIndicatorStyle lipgloss.Style
-}
-
-// DefaultStyles returns a set of default style definitions for this tree
-// component.
-func DefaultStyles() (s Styles) {
-	verySubduedColor := lipgloss.AdaptiveColor{Light: "#DDDADA", Dark: "#3C3C3C"}
-	subduedColor := lipgloss.AdaptiveColor{Light: "#9B9B9B", Dark: "#5C5C5C"}
-
-	s.TreeStyle = lipgloss.NewStyle()
-	s.HelpStyle = lipgloss.NewStyle().PaddingTop(1)
-
-	s.NodeStyle = lipgloss.NewStyle().Foreground(lipgloss.AdaptiveColor{Light: "#9B9B9B", Dark: "#b0b0b0"})
-	s.nodeFunc = func(_ Nodes, _ int) lipgloss.Style {
-		return s.NodeStyle
-	}
-
-	s.SelectedNodeStyle = s.NodeStyle.Foreground(lipgloss.Color("212")).Bold(true)
-	s.selectedNodeFunc = func(_ Nodes, _ int) lipgloss.Style {
-		return s.SelectedNodeStyle
-	}
-
-	s.RootNodeStyle = lipgloss.NewStyle().
-		Foreground(lipgloss.AdaptiveColor{Light: "#EE6FF8", Dark: "#EE6FF8"})
-	s.rootNodeFunc = func(_ Nodes, _ int) lipgloss.Style {
-		return s.RootNodeStyle
-	}
-
-	s.ParentNodeStyle = lipgloss.NewStyle().Foreground(lipgloss.Color("99"))
-	s.parentNodeFunc = func(_ Nodes, _ int) lipgloss.Style {
-		return s.ParentNodeStyle
-	}
-
-	s.CursorStyle = lipgloss.NewStyle().PaddingRight(1).Foreground(lipgloss.Color("212")).Bold(true)
-
-	s.EnumeratorStyle = lipgloss.NewStyle().Foreground(verySubduedColor).PaddingRight(1)
-	s.OpenIndicatorStyle = lipgloss.NewStyle().Foreground(subduedColor)
-
-	return s
-}
-
 const spacebar = " "
 
 // KeyMap is the key bindings for different actions within the tree.
@@ -562,26 +493,21 @@ func (m Model) helpView() string {
 	return m.styles.HelpStyle.Render(m.Help.View(m))
 }
 
+// Root returns the root node of the tree.
+// Equivalent to calling `Model.Node(0)`
+func (m *Model) Root() *Node {
+	return m.root
+}
+
 // FlatNodes returns all items in the tree as a flat list.
-func (m *Model) FlatNodes() []*Node {
-	return m.root.FlatNodes()
+func (m *Model) AllNodes() []*Node {
+	return m.root.AllNodes()
 }
 
 func (m *Model) setAttributes() {
 	setDepths(m.root, 0)
 	setSizes(m.root)
 	setYOffsets(m.root)
-}
-
-// FlatNodes returns all descendant items in as a flat list.
-func (t *Node) FlatNodes() []*Node {
-	res := []*Node{t}
-	children := t.tree.Children()
-	for i := 0; i < children.Length(); i++ {
-		child := children.At(i)
-		res = append(res, child.(*Node).FlatNodes()...)
-	}
-	return res
 }
 
 // setSizes updates each Node's size
@@ -646,15 +572,19 @@ func (m *Model) updateStyles() {
 		m.root.RootStyle(m.rootStyle())
 	}
 
-	items := m.FlatNodes()
-	opts := &itemOptions{
+	items := m.AllNodes()
+	opts := m.getItemOpts()
+	for _, item := range items {
+		item.opts = *opts
+	}
+}
+
+func (m *Model) getItemOpts() *itemOptions {
+	return &itemOptions{
 		openCharacter:   m.OpenCharacter,
 		closedCharacter: m.ClosedCharacter,
 		treeYOffset:     m.yOffset,
-		styles:          &m.styles,
-	}
-	for _, item := range items {
-		item.opts = opts
+		styles:          m.styles,
 	}
 }
